apps/api/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio>=4.0.0
asyncpg==0.30.0
boto3==1.40.71
botocore==1.40.71
click==8.3.0
fastapi==0.121.1
h11==0.16.0
idna==3.11
jmespath==1.0.1
pydantic==2.12.4
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
starlette>=0.37
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
reportlab==4.4.4
camelot-py==0.11.0
python-dotenv==1.0.1
httpx==0.27.0
pdfplumber==0.10.3
PyPDF2==3.0.1
pdfminer.six==20221105
pillow==10.2.0
opencv-python-headless>=4.7.0.68
openai>=1.3.7
